// web/lib/api/generated/model/branding.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export interface Branding {
  /** @maxLength 120 */
  display_name: string;
  /** @maxLength 7 */
  primary_color: string;
  /** @nullable */
  readonly logo_url: string | null;
  /** @nullable */
  readonly favicon_url: string | null;
  /** @nullable */
  readonly app_icon_url: string | null;
}